Move elaboration into its own module.
