+++ complexity-analysis.cabal
@@ -12,7 +12,8 @@ build-type:          Simple
 cabal-version:       >=1.10
 
 library
-  exposed-modules:     Control.Monad.Fresh
+  exposed-modules:     Analysis.Elaboration
+                     , Control.Monad.Fresh
                      , Data.Complexity
                      , Data.Env
                      , Data.Expr
